🧹 [code health improvement] Refactor DataGravityService to improve readability
🎯 What: Extracted DataGravityService into smaller, single-responsibility modules (types, analyzer, exporter, and insights) inside a new data-gravity directory. Re-exported them from the original service file.
💡 Why: The original file was over 500 lines long and contained multiple distinct responsibilities. Breaking it down makes the code easier to maintain, test, and understand.
✅ Verification: Ran build and test suites to verify that the refactoring did not break existing functionality or types. Confirmed via codebase review that public API exports match.
✨ Result: improved code structure and separated concerns into logical pieces without behavior changes.
